Part 3: User Create and Login

+++ api/models/todoModel.js
@@ -13,11 +13,12 @@ async function getDbServiceInstance() {
 }
 
 //Get all
-async function getAllTodosModel(){
+async function getAllTodosByIdModel(userId){
         try{
             const dbServiceInstance = await getDbServiceInstance();
-            const query = 'SELECT * FROM public.todolist';
-            const response = await dbServiceInstance.queryMethod(query);
+            const query = 'SELECT * FROM public.todolist WHERE user_id = $1';
+            const params = [userId];
+            const response = await dbServiceInstance.queryMethod(query, params);
             return response;
         } catch (error){
             throw error;
@@ -25,11 +26,11 @@ async function getAllTodosModel(){
 }
 
 //Add new
-async function addTodoModel(newTodo){
+async function addTodoModel(newTodo, userId){
     try{
         const dbServiceInstance = await getDbServiceInstance();
-        const query = 'INSERT INTO public.todolist (title) VALUES ($1) RETURNING *';
-        const params = [newTodo];
+        const query = 'INSERT INTO public.todolist (title, user_id) VALUES ($1, $2) RETURNING *';
+        const params = [newTodo, userId];
         const response = await dbServiceInstance.queryMethod(query, params);
         return response[0];
     }catch(error){
@@ -38,23 +39,22 @@ async function addTodoModel(newTodo){
 }
 
 //Toggler
-async function toggleTodoModel(id) {
+async function toggleTodoModel(id, userId) {
     try{
-        console.log("Model Called");
         const dbServiceInstance = await getDbServiceInstance();
-        const selectQuery = 'SELECT completed FROM public.todolist WHERE todolist_id = $1';
-        const selectParams = [id];
-        console.log(selectParams);
+        const selectQuery = 'SELECT completed FROM public.todolist WHERE todolist_id = $1 AND user_id = $2';
+        const selectParams = [id, userId];
+
         const selectResponse = await dbServiceInstance.queryMethod(selectQuery, selectParams);
-        console.log(selectResponse);
+
         let check = selectResponse[0].completed;
-        console.log("pre",check);
+
         if (check){check = false;}else{check = true};
-        console.log("post",check);
-        const updateQuery = 'UPDATE public.todolist SET completed = $1 WHERE todolist_id = $2 RETURNING *';
-        const updateParams = [check, id];
+
+        const updateQuery = 'UPDATE public.todolist SET completed = $1 WHERE todolist_id = $2 AND user_id = $3 RETURNING *';
+        const updateParams = [check, id, userId];
         const updateResponse = await dbServiceInstance.queryMethod(updateQuery, updateParams);
-        console.log(updateResponse);
+
         return updateResponse;
     }catch(error){
         throw error;
@@ -62,11 +62,11 @@ async function toggleTodoModel(id) {
 }
 
 //Remove
-async function removeTodoModel(id){
+async function removeTodoModel(id, userId){
     try{
         const dbServiceInstance = await getDbServiceInstance();
-        const query = 'DELETE FROM public.todolist WHERE todolist_id = $1 RETURNING *';
-        const params = [id.id];
+        const query = 'DELETE FROM public.todolist WHERE todolist_id = $1 AND user_id = $2 RETURNING *';
+        const params = [id, userId];
         const response = await dbServiceInstance.queryMethod(query, params);
         return response;
     }catch(error){
@@ -75,11 +75,11 @@ async function removeTodoModel(id){
 }
 
 //Update Title
-async function updateTodoModel(id, updatedFields){
+async function updateTodoModel(id, updatedFields, userId ){
     try{
         const dbServiceInstance = await getDbServiceInstance();
-        const query = 'UPDATE public.todolist SET title = $2 WHERE todolist_id = $1 RETURNING *';
-        const params = [id.id, updatedFields];
+        const query = 'UPDATE public.todolist SET title = $2 WHERE todolist_id = $1 AND user_id = $3 RETURNING *';
+        const params = [id, updatedFields, userId];
         const response = await dbServiceInstance.queryMethod(query,params);
         return response;
     }catch(error){
@@ -89,7 +89,7 @@ async function updateTodoModel(id, updatedFields){
 
 
 module.exports = {
-    getAllTodosModel,
+    getAllTodosByIdModel,
     addTodoModel,
     toggleTodoModel,
     removeTodoModel,
